Completed Order Form

+++ Assignment4/Modules/product.cs
@@ -11,7 +11,7 @@ namespace Assignment4.Modules
         public short productID { get; set; }
 
         [Column(TypeName = "money")]
-        public decimal? cost { get; set; }
+        public double? cost { get; set; }
 
         [StringLength(255)]
         public string manufacturer { get; set; }
